Refactor splitting files
